Refactored creature and map scripts into 'game-content' folder.
