+++ src/engine/sensors/facing-sensor.ts
