+++ src/engine/sensors/startle-sensor.ts
