+++ src/engine/sensors/tile-visibility-sensor.ts
