+++ src/game-content/scripts/creatures/player.ts
