+++ src/game-content/scripts/quests/problem-of-scale.ts
