+++ src/ui/components/map-scene-graph.ts
@@ -2,7 +2,7 @@ import { Creature } from 'engine/creature'
 import { CellCoordinate, MapCell } from 'engine/map/map'
 import { MapSymbol } from 'engine/map/map-symbol'
 import { getCreatureSymbol, getItemSymbol, getTerrainSymbol, withDefaultBackground } from 'engine/map/map-symbolizer'
-import { isTileVisibleTo } from 'engine/scripts/tile-visibility-sensor'
+import { isTileVisibleTo } from 'engine/sensors/tile-visibility-sensor'
 import { World } from 'engine/world'
 import { compact, forEach, get, keys, map as lodashMap, values, without } from 'lodash/fp'
 import * as PIXI from 'pixi.js'
